removed BinFetchingProtocol
