+++ Sources/MercadoPagoCheckout/Domain/CheckoutServiceProtocol.swift
@@ -8,6 +8,9 @@ import CoreMethods
 
 protocol CheckoutServiceProtocol: Sendable {
     func identificationTypes() async throws -> [IdentificationType]
+    func paymentMethod(bin: String) async throws -> [PaymentMethod]
+    func issuers(bin: String, paymentMethodID: String) async throws -> [Issuer]
+    func installments(amount: Double, bin: String) async throws -> [Installment]
     func fetchBinData(
         bin: String,
         amount: Double?,
